Updated datalake mock cases into actions file
